requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
blinker==1.7.0
cachetools==5.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.87
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
distinctipy==1.3.4
distlib==0.3.9
Django==5.0.4
djangorestframework==3.15.1
dm-tree==0.1.8
docstring_parser==0.16
etils==1.9.4
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
flax==0.9.0
fonttools==4.51.0
fsspec==2024.9.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
gradio==5.6.0
gradio_client==1.4.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
humanize==4.10.0
idna==3.7
immutabledict==4.2.1
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.1.2
jax==0.4.33
jaxlib==0.4.33
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
narwhals==1.14.2
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optax==0.2.3
optree==0.11.0
orbax-checkpoint==0.6.4
orjson==3.10.12
packaging==23.2
panda==0.3.1
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
plotly==5.24.1
promise==2.3
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
pure-eval==0.2.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.12
python-version==0.0.2
pytz==2024.1
pywebio==1.8.3
pywin32==306
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.21.0
rsa==4.9
ruff==0.8.0
safehttpx==0.1.1
scikit-learn==1.5.2
scipy==1.13.0
seaborn==0.13.2
segment-anything==1.0
semantic-version==2.10.0
shellingham==1.5.4
simple-parsing==0.1.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.41.3
streamlit==1.40.2
streamlit-drawable-canvas==0.9.3
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.1
tensorflow-datasets==4.9.7
tensorflow-decision-forests==1.8.1
tensorflow-estimator==2.15.0
tensorflow-gpu==2.10.1
tensorflow-hub==0.16.1
tensorflow-intel==2.15.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.16.1
tensorflowjs==4.21.0
tensorstore==0.1.66
termcolor==2.4.0
tf_keras==2.15.1
threadpoolctl==3.5.0
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
urllib3==2.2.1
user-agents==2.2.0
uvicorn==0.32.1
virtualenv==20.29.2
watchdog==6.0.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.2
wrapt==1.14.1
wurlitzer==3.1.1
zipp==3.20.2
